Change telemetry references to open telemetry. Fix wording
