+++ .github/workflows/tests.yml
@@ -25,7 +25,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install wheel
-        pip install -e .[subscriptions,telemetry,dev]
+        pip install -e .[subscriptions,opentelemetry,dev]
     - name: Pytest
       run: |
         pytest
