+++ ariadne_codegen/settings.py
@@ -8,13 +8,13 @@ from typing import Dict, List
 
 from .client_generators.constants import (
     DEFAULT_ASYNC_BASE_CLIENT_NAME,
+    DEFAULT_ASYNC_BASE_CLIENT_OPEN_TELEMETRY_NAME,
+    DEFAULT_ASYNC_BASE_CLIENT_OPEN_TELEMETRY_PATH,
     DEFAULT_ASYNC_BASE_CLIENT_PATH,
-    DEFAULT_ASYNC_BASE_CLIENT_WITH_TELEMETRY_NAME,
-    DEFAULT_ASYNC_BASE_CLIENT_WITH_TELEMETRY_PATH,
     DEFAULT_BASE_CLIENT_NAME,
+    DEFAULT_BASE_CLIENT_OPEN_TELEMETRY_NAME,
+    DEFAULT_BASE_CLIENT_OPEN_TELEMETRY_PATH,
     DEFAULT_BASE_CLIENT_PATH,
-    DEFAULT_BASE_CLIENT_WITH_TELEMETRY_NAME,
-    DEFAULT_BASE_CLIENT_WITH_TELEMETRY_PATH,
 )
 from .client_generators.scalars import ScalarData
 from .exceptions import InvalidConfiguration
@@ -66,7 +66,7 @@ class ClientSettings(BaseSettings):
     include_comments: CommentsStrategy = field(default=CommentsStrategy.STABLE)
     convert_to_snake_case: bool = True
     async_client: bool = True
-    telemetry_client: bool = False
+    opentelemetry_client: bool = False
     files_to_include: List[str] = field(default_factory=list)
     scalars: Dict[str, ScalarData] = field(default_factory=dict)
 
@@ -112,16 +112,16 @@ class ClientSettings(BaseSettings):
     def _set_default_base_client_data(self):
         default_clients_map = {
             (True, True): (
-                DEFAULT_ASYNC_BASE_CLIENT_WITH_TELEMETRY_PATH,
-                DEFAULT_ASYNC_BASE_CLIENT_WITH_TELEMETRY_NAME,
+                DEFAULT_ASYNC_BASE_CLIENT_OPEN_TELEMETRY_PATH,
+                DEFAULT_ASYNC_BASE_CLIENT_OPEN_TELEMETRY_NAME,
             ),
             (True, False): (
                 DEFAULT_ASYNC_BASE_CLIENT_PATH,
                 DEFAULT_ASYNC_BASE_CLIENT_NAME,
             ),
             (False, True): (
-                DEFAULT_BASE_CLIENT_WITH_TELEMETRY_PATH,
-                DEFAULT_BASE_CLIENT_WITH_TELEMETRY_NAME,
+                DEFAULT_BASE_CLIENT_OPEN_TELEMETRY_PATH,
+                DEFAULT_BASE_CLIENT_OPEN_TELEMETRY_NAME,
             ),
             (False, False): (
                 DEFAULT_BASE_CLIENT_PATH,
@@ -129,7 +129,9 @@ class ClientSettings(BaseSettings):
             ),
         }
         if not self.base_client_name and not self.base_client_file_path:
-            path, name = default_clients_map[(self.async_client, self.telemetry_client)]
+            path, name = default_clients_map[
+                (self.async_client, self.opentelemetry_client)
+            ]
             self.base_client_name = name
             self.base_client_file_path = path.as_posix()
