+++ pyproject.toml
@@ -45,7 +45,7 @@ dev = [
   "requests-toolbelt",
 ]
 subscriptions = ["websockets~=11.0"]
-telemetry = ["opentelemetry-api"]
+opentelemetry = ["opentelemetry-api"]
 
 [project.scripts]
 ariadne-codegen = "ariadne_codegen.main:main"
